Get alumnos ended
