+++ .flaskenv
@@ -1,2 +1,2 @@
 FLASK_APP=app
-FLASK_DEBUG=True
+FLASK_DEBUG=1
