+++ Procfile
@@ -1,0 +1,2 @@
+# Modify this Procfile to fit your needs
+web: gunicorn server:app
