+++ fly.toml
@@ -1,0 +1,28 @@
+# fly.toml file generated for integrador-flask-api on 2023-10-12T12:19:15-03:00
+
+app = "integrador-flask-api"
+primary_region = "scl"
+
+[build]
+  builder = "paketobuildpacks/builder:base"
+
+[env]
+  PORT = "8080"
+  DB_HOST= "silly.db.elephantsql.com"
+  DB_NAME= "wcfjjmav"
+  DB_USER_NAME= "wcfjjmav"
+  DB_PASSWORD= "0WMsyEdub8wgpfSR-3OCHQIxiuy_TY5O"
+  DB_URI= "postgresql://wcfjjmav:0WMsyEdub8wgpfSR-3OCHQIxiuy_TY5O@silly.db.elephantsql.com:5432/wcfjjmav"
+  FLASK_APP="app"
+  FLASK_DEBUG="1"
+
+[http_service]
+  internal_port = 8080
+  force_https = true
+
+[checks]
+  [checks.alive]
+    type = "tcp"
+    interval = "15s"
+    timeout = "2s"
+    grace_period = "5s"
